**Update 12.11.2024**
We created a fork of the original code, removing portions of the initial code that were irrelevant to our project. Now, we need to access the disk containing the modified code to upload it into this fork.

Additionally, we added a `requirements` file, which should be installed using `pip` with the `--no-dependencies` flag; otherwise, it will result in an error.